--- frontend/src/components/pages/projects/PastProject.js
@@ -1,16 +1,0 @@
-import React from "react";
-import Header from "../../header";
-import Footer from "../../footer";
-import PastProjects from "../../projectPageComponents/pastProjects/PastProjects.component";
-
-const Projects = () => {
-  return (
-    <div>
-      <Header />
-      <PastProjects />
-      <Footer />
-    </div>
-  );
-};
-
-export default Projects;
