--- frontend/src/components/pages/projects/Projects.js
@@ -1,16 +1,0 @@
-import React from "react";
-import Header from "../../header";
-import Footer from "../../footer";
-import ProjectHistory from "../../projectPageComponents/ProjectPage.components";
-
-const Projects = () => {
-  return (
-    <div>
-      <Header />
-      <ProjectHistory />
-      <Footer />
-    </div>
-  );
-};
-
-export default Projects;
