+++ frontend/src/components/projectPageComponents/ProjectPage.components.jsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import Card from "@mui/material/Card";
 import Box from "@mui/material/Box";
 import { Link } from "react-router-dom";
-import "../pages/projects/projects.css";
+import "../../pages/projects/projects.css";
 
 const squareCardStyles = {
   width: "500px", // Set the width of the square card
