+++ frontend/src/components/projectPageComponents/pastProjects/PastProjects.component.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { getProjects } from "../../api/api.components";
 import PastProjectList from "../projectList/ProjectsList.components";
 import SearchBox from "../projectList/SearchBox.components";
-import "../../pages/projects/projects.css";
+import "../../../pages/projects/projects.css";
 
 let PastProjects = () => {
   const [projects, setProjects] = useState([]);
