+++ frontend/src/components/projectPageComponents/projectList/ProjectsList.components.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ProjectCard from "./projectCard.components"; // Update the import path to match your component's location
 import "./project-list.styles.css";
-import "../../pages/projects/projects.css";
+import "../../../pages/projects/projects.css";
 
 const ProjectList = ({ projects }) => {
   return (
