+++ frontend/src/pages/projects/CurrentProjects.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "../../header";
-import Footer from "../../footer";
+import Header from "../../components/header";
+import Footer from "../../components/footer";
 
 const Projects = () => {
   return (
