+++ frontend/src/pages/projects/PastProject.js
@@ -1,0 +1,16 @@
+import React from "react";
+import Header from "../../components/header";
+import Footer from "../../components/footer";
+import PastProjects from "../../components/projectPageComponents/pastProjects/PastProjects.component";
+
+const Projects = () => {
+  return (
+    <div>
+      <Header />
+      <PastProjects />
+      <Footer />
+    </div>
+  );
+};
+
+export default Projects;
