+++ frontend/src/pages/projects/Projects.js
@@ -1,0 +1,16 @@
+import React from "react";
+import Header from "../../components/header";
+import Footer from "../../components/footer";
+import ProjectPage from "../../components/projectPageComponents/ProjectPage.components";
+
+const Projects = () => {
+  return (
+    <div>
+      <Header />
+      <ProjectPage />
+      <Footer />
+    </div>
+  );
+};
+
+export default Projects;
